This revision requires GameDir 1873 to work because of a new icon in ItemInfoAdvancedIcons.sti!!!
- Weapons.xml tags <bRecoilX> and <bRecoilY> are now FLOAT type values. This allows to differentiate weapons by smaller steps.
- Items.xml tags <RecoilModifierX> and <RecoilModifierY> are now FLOAT types as well.
- Items.xml tag <PercentRecoilModifier> will now be displayed in UDB advanced tab.
- Item_Settings.ini has new modifiers for gun type specific <bRecoilX> and <bRecoilY>. This allows modders to quickly adjust recoil values for any gun type.
- fixed UDB description for <RecoilModifierX> and <RecoilModifierY>
- All recoil calculations now use FLOAT values for more precision.


git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@6613 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Item Types.h
@@ -1099,8 +1099,8 @@ typedef struct
 	INT16 aimlevelsmodifier[3];
 	// HEADROCK HAM 4: New modifiers that do not require a stance array, since they affect the gun objectively, not
 	// subjectively.
-	INT16 RecoilModifierX;
-	INT16 RecoilModifierY;
+	FLOAT RecoilModifierX;
+	FLOAT RecoilModifierY;
 	INT16 PercentRecoilModifier;
 	INT16 percentaccuracymodifier;
 	FLOAT scopemagfactor;
